+++ src/PersistentModel.php
@@ -4,6 +4,7 @@ namespace Rubix\ML;
 
 use Rubix\ML\Datasets\Dataset;
 use Rubix\ML\Persisters\Persister;
+use Rubix\ML\Other\Traits\ProbaSingle;
 use Rubix\ML\Other\Traits\PredictsSingle;
 use InvalidArgumentException;
 use RuntimeException;
@@ -20,7 +21,7 @@ use RuntimeException;
  */
 class PersistentModel implements Estimator, Learner, Wrapper, Probabilistic
 {
-    use PredictsSingle;
+    use PredictsSingle, ProbaSingle;
     
     /**
      * The underlying persistable estimator instance.